reserve

+++ src/main/java/charger/main/persistence/EvCarsRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import charger.main.domain.EvCars;
 
-public interface EvCarsRepository extends JpaRepository<EvCars, Integer>{
+public interface EvCarsRepository extends JpaRepository<EvCars, Long>{
 	@Query("select ec from EvCars ec where ec.carId = :carId")
-	EvCars getByCarId(Integer carId);
+	EvCars getByCarId(Long carId);
 	
 	List<EvCars> findByBrand(String brand);
 }
